Blocklist added

+++ code/resources/item.py
@@ -1,7 +1,9 @@
-import re
-import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt,
+    get_jwt_identity
+)
 from models.item import ItemModel
 
 
@@ -26,7 +28,7 @@ class Item(Resource):
         else:
             return {'message': 'Item not found'}, 404
 
-
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': f'An item with name {name} already exists'}, 400
@@ -42,8 +44,11 @@ class Item(Resource):
 
         return item.json(), 201
 
-
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -67,5 +72,13 @@ class Item(Resource):
 
 
 class Items(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+            }, 200
